fix: clean up dead code and leftovers
